Bumped all integer tags to int64 to avoid data loss

+++ Source/PCGExtendedToolkit/Private/Data/PCGExPointIO.cpp
@@ -880,7 +880,7 @@ for (int i = 0; i < ReducedNum; i++){Range[i] = Range[InIndices[i]];}}
 
 	bool FPointIOTaggedDictionary::CreateKey(const TSharedRef<FPointIO>& PointIOKey)
 	{
-		PCGExCommon::DataIDType TagValue = PointIOKey->Tags->GetOrSet<int32>(TagIdentifier, PointIOKey->GetInOut()->GetUniqueID());
+		PCGExCommon::DataIDType TagValue = PointIOKey->Tags->GetOrSet<int64>(TagIdentifier, PointIOKey->GetInOut()->GetUniqueID());
 		if (TagMap.Contains(TagValue->Value)) { return false; }
 		TagMap.Add(TagValue->Value, Entries.Add(MakeShared<FPointIOTaggedEntries>(PointIOKey, TagIdentifier, TagValue)));
 		return true;
@@ -888,7 +888,7 @@ for (int i = 0; i < ReducedNum; i++){Range[i] = Range[InIndices[i]];}}
 
 	bool FPointIOTaggedDictionary::RemoveKey(const TSharedRef<FPointIO>& PointIOKey)
 	{
-		const PCGExCommon::DataIDType TagValue = PointIOKey->Tags->GetTypedValue<int32>(TagIdentifier);
+		const PCGExCommon::DataIDType TagValue = PointIOKey->Tags->GetTypedValue<int64>(TagIdentifier);
 
 		if (!TagValue) { return false; }
 
@@ -902,7 +902,7 @@ for (int i = 0; i < ReducedNum; i++){Range[i] = Range[InIndices[i]];}}
 
 	bool FPointIOTaggedDictionary::TryAddEntry(const TSharedRef<FPointIO>& PointIOEntry)
 	{
-		const PCGExCommon::DataIDType TagValue = PointIOEntry->Tags->GetTypedValue<int32>(TagIdentifier);
+		const PCGExCommon::DataIDType TagValue = PointIOEntry->Tags->GetTypedValue<int64>(TagIdentifier);
 		if (!TagValue) { return false; }
 
 		if (const int32* Index = TagMap.Find(TagValue->Value))
